+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Dependabot: Bump http-cache-semantics from 4.1.0 to 4.1.1
+
 ## [0.5.0] - 2023-01-25
 
 ### Added